GROMACS version 2022

+++ src/gromacs/nbnxm/pairsearch.h
@@ -1,11 +1,9 @@
 /*
  * This file is part of the GROMACS molecular simulation package.
  *
- * Copyright (c) 2012,2013,2014,2015,2016 by the GROMACS development team.
- * Copyright (c) 2017,2018,2019,2020, by the GROMACS development team, led by
- * Mark Abraham, David van der Spoel, Berk Hess, and Erik Lindahl,
- * and including many others, as listed in the AUTHORS file in the
- * top-level source directory and at http://www.gromacs.org.
+ * Copyright 2012- The GROMACS Authors
+ * and the project initiators Erik Lindahl, Berk Hess and David van der Spoel.
+ * Consult the AUTHORS/COPYING files and https://www.gromacs.org for details.
  *
  * GROMACS is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public License
@@ -19,7 +17,7 @@
  *
  * You should have received a copy of the GNU Lesser General Public
  * License along with GROMACS; if not, see
- * http://www.gnu.org/licenses, or write to the Free Software Foundation,
+ * https://www.gnu.org/licenses, or write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA.
  *
  * If you want to redistribute modifications to GROMACS, please
@@ -28,10 +26,10 @@
  * consider code for inclusion in the official distribution, but
  * derived work must not be called official GROMACS. Details are found
  * in the README & COPYING files - if they are missing, get the
- * official version at http://www.gromacs.org.
+ * official version at https://www.gromacs.org.
  *
  * To help us fund GROMACS development, we humbly ask that you cite
- * the research papers on the package. Check out http://www.gromacs.org.
+ * the research papers on the package. Check out https://www.gromacs.org.
  */
 
 /*! \internal \file
@@ -40,7 +38,7 @@
  *
  * The PairSearch class holds the domain setup, the search grids
  * and helper object for the pair search. It manages the search work.
- * The actual gridding and pairlist generation is performeed by the
+ * The actual gridding and pairlist generation is performed by the
  * GridSet/Grid and PairlistSet/Pairlist classes, respectively.
  *
  * \author Berk Hess <hess@kth.se>
@@ -54,7 +52,6 @@
 #include <memory>
 #include <vector>
 
-#include "gromacs/domdec/domdec.h"
 #include "gromacs/math/vectypes.h"
 #include "gromacs/nbnxm/atomdata.h"
 #include "gromacs/timing/cyclecounter.h"
@@ -151,7 +148,7 @@ struct PairsearchWork
 
     ~PairsearchWork();
 
-    //! Buffer to avoid cache polution
+    //! Buffer to avoid cache pollution
     gmx_cache_protect_t cp0;
 
     //! Temporary buffer for sorting atoms within a grid column
@@ -186,7 +183,7 @@ public:
                    const gmx::UpdateGroupsCog*    updateGroupsCog,
                    gmx::Range<int>                atomRange,
                    real                           atomDensity,
-                   gmx::ArrayRef<const int>       atomInfo,
+                   gmx::ArrayRef<const int64_t>   atomInfo,
                    gmx::ArrayRef<const gmx::RVec> x,
                    int                            numAtomsMoved,
                    const int*                     move,
@@ -194,8 +191,18 @@ public:
     {
         cycleCounting_.start(enbsCCgrid);
 
-        gridSet_.putOnGrid(box, ddZone, lowerCorner, upperCorner, updateGroupsCog, atomRange,
-                           atomDensity, atomInfo, x, numAtomsMoved, move, nbat);
+        gridSet_.putOnGrid(box,
+                           ddZone,
+                           lowerCorner,
+                           upperCorner,
+                           updateGroupsCog,
+                           atomRange,
+                           atomDensity,
+                           atomInfo,
+                           x,
+                           numAtomsMoved,
+                           move,
+                           nbat);
 
         cycleCounting_.stop(enbsCCgrid);
     }
